“Recommender” Stored Cross-Site Scripting (XSS)
The Recommender feature of the LMS allows students to share links and references with each other, but it also vulnerable to stored XSS attacks. Stored XSS payloads will be served up to all visitors of the Recommender page, allowing session hijacking and PII leaks. Any student is capable of performing this attack and targeting all other students, staff members, or administrators; targeting of a staff member or administrator would allow a student to escalate privilege levels.

The XSS payload is delivered to client browsers with HTML entities and quote characters encoded. It is part of a larger body of encoded HTML which is decoded by JavaScript and dynamically added to the page. This decoding allows the malicious JavaScript to run, and may require rearchitecting data delivery to prevent.

EDUCATOR-3207